[jimple] General Archive Detection (#4460)
Currently, archives are determined by extension, and then corruption is detected if there is a failure along the line.

This change ignores file name completely, and runs through the whole zip to make sure it's valid and uncorrupted before processing. This adds a bit more computation but helps with validity.