Dependency Housekeeping: Removed `circe`, `jackson` & Organised Versions (#4472)
Joern has **a lot** of dependencies, and I've found `meta-build` out of sync with some `build.sbt` files specifying the same dependencies.

Also, `Versions.scala` is used haphazardly and is inconsistent. This change congregates all versions there, so we can see all versions of dependencies in Joern.

This all started by wondering why we have `jackson-databind`, `ujson`, and `circe` to deserialize JSON, so I removed `circe` and `jackson-databind` while keeping `ujson`. In any-case, this organizes dependencies and paves the way for removing unused ones.

Note, some code in `jssrc2cpg` (@max-leuthaeuser) and `gosrc2cpg` (@pandurangpatil) were modified for the migration to deserialize with `ujson` instead of the existing libraries.