[python] Correct `try-except` AST `order` Property and Structure (#4579)
The current AST of the try except does not keep the `except` bodies as `order=2` and pass empty block nodes for `finally` and `else` bodies if they are not there.

This change keeps the `except` bodies as `order=2` and groups the `finally` and `else` bodies under a single body, while discarding empty blocks if they are unimplemented.

Partially addresses #4578, as the node is unfortunately still dangling. This could also be because `file` is a built-in function call and is re-assigned in `a9_lab` from the initial assignment under the parent `<module>`.

One note is that setting the source to `cpg.parameter.nameExact("request").l` does not work, and it may be due to the excessive AST depth under the assignment to `file`.