[ruby] Grouped `facter` Bug Fixes (#4585)
The [`facter`](https://github.com/puppetlabs/facter) repository causes this frontend to throw a lot of warnings. This PR fixes much of the low-hanging fruit, and improves the warning techniques to help make the warnings more meaningful. There are still, however, quite a few syntax errors to handle from this repository.

* Handles regexes generated from string formating.
* Handled the `!~` regex not match function. Long term any unrecognized functions are interpreted as calls.
* Added more cases for implicit returns