[gosrc2cpg] Partial gosrc2cpg frontend perforamnce optimisations (#4668)
* Partial gosrc2cpg frontend perforamnce optimisations

1. Changed parsing of AST json inside AST Creator itself.
2. Made changes in download dependency processing to generate AST only
for the used packages and subsequently only processing used packages.
There was a change required inside `goastgen` utility to support the
`-include` input option.

TODO:
At this moment `AstCreator` gets instantiated while first-level
processing for building the cache and it stays in memory till
`AstCreationPass` is done. In subsequent changes, we will split the
processing in such a way that we create `AstCreator` for that pass and
destroy it once it's used. Instantiate it again for `AstCreationPass`.

The above change is done as part of the larger change.

* small pending change

* Refactored some code to use meaningful name in the context